# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Learning Platform API",
        "description": "API for Learning Platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/quizzes/{quizId}/questions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Get questions for a quiz",
                "operationId": "8a042cb88af67edc55989f50b8f628eb",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of questions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Question"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Create a new question",
                "operationId": "5c9f901cd140f565401fd30590b30dbe",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question_text",
                                    "question_type",
                                    "points",
                                    "correct_answer"
                                ],
                                "properties": {
                                    "question_text": {
                                        "type": "string",
                                        "example": "What is Vue.js?"
                                    },
                                    "question_type": {
                                        "type": "string",
                                        "enum": [
                                            "MULTIPLE_CHOICE",
                                            "TRUE_FALSE",
                                            "SHORT_ANSWER",
                                            "ESSAY"
                                        ]
                                    },
                                    "points": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_index": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "options": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "correct_answer": {
                                        "type": "string"
                                    },
                                    "explanation": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/questions/{questionId}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Get question details",
                "operationId": "edfc0c72ca1e9f11e6c94e16cd6d5daa",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Questions"
                ],
                "summary": "Update question",
                "operationId": "e2aa92271e8d907e73fcbedc0577bce5",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "question_text": {
                                        "type": "string"
                                    },
                                    "question_type": {
                                        "type": "string",
                                        "enum": [
                                            "MULTIPLE_CHOICE",
                                            "TRUE_FALSE",
                                            "SHORT_ANSWER",
                                            "ESSAY"
                                        ]
                                    },
                                    "points": {
                                        "type": "integer"
                                    },
                                    "order_index": {
                                        "type": "integer"
                                    },
                                    "options": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "correct_answer": {
                                        "type": "string"
                                    },
                                    "explanation": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Question updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Delete question",
                "operationId": "7bad0f55b1e4fba994977337462e153f",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Question": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "question_text": {
                        "type": "string"
                    },
                    "question_type": {
                        "type": "string",
                        "enum": [
                            "MULTIPLE_CHOICE",
                            "TRUE_FALSE",
                            "SHORT_ANSWER",
                            "ESSAY"
                        ]
                    },
                    "points": {
                        "type": "integer",
                        "default": 1
                    },
                    "order_index": {
                        "type": "integer"
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "correct_answer": {
                        "type": "string"
                    },
                    "explanation": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token (without 'Bearer ' prefix)",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}